First draft of colorTrainingDataAnalysis
At the moment, the code is still a bit buggy and thrown together. The overall purpose of this code is to take my reaction time for identifying each square on a chess board and help me visualize where I need to get better. It delivers a bar graph with my reaction time grouped by chess board file. Ex. 'a', 'b', 'f', ect.